# index.js
const express = require('express');
const bodyParser = require('body-parser');
const app = express();
const port = 3000;

app.use(bodyParser.json());

let todos = [];

// Routes
// 1. Get all todos
app.get('/todos', (req, res) => {
  res.json(todos);
});

// 2. Add a new todo
app.post('/todos', (req, res) => {
  const { task } = req.body;
  if (!task) {
    return res.status(400).json({ error: 'Task is required' });
  }
  const newTodo = { id: todos.length + 1, task, completed: false };
  todos.push(newTodo);
  res.status(201).json(newTodo);
});

// 3. Update a todo
// app.put('/todos/:id', (req, res) => {
//   const { id } = req.params;
//   const { task, completed } = req.body;

//   const todo = todos.find((t) => t.id === parseInt(id));
//   if (!todo) {
//     return res.status(404).json({ error: 'Todo not found' });
//   }

//   if (task !== undefined) todo.task = task;
//   if (completed !== undefined) todo.completed = completed;

//   res.json(todo);
// });

// 4. Delete a todo
app.delete('/todos/danger/:id', (req, res) => {
  const { id } = req.params;

  const index = todos.findIndex((t) => t.id === parseInt(id));
  if (index === -1) {
    return res.status(404).json({ error: 'Todo not found' });
  }

  const deletedTodo = todos.splice(index, 1);
  res.json(deletedTodo[0]);
});

// Start server
app.listen(port, () => {
  console.log(`🚀 App running at http://localhost:${port}`);
});